move gas costs to its own file
